requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
yt-dlp==2023.9.24
lxml==4.9.3
python-dotenv==1.0.0
gunicorn==21.2.0